modified:   allegro.log 	deleted:    assets/image/floor.png 	deleted:    assets/image/projectile.png 	deleted:    assets/image/source.txt 	deleted:    assets/image/teleport.png 	deleted:    assets/image/tree.png 	renamed:    assets/image/white_house_background.png -> assets/map/white_house_background.png 	renamed:    assets/image/white_house_ground.png -> assets/map/white_house_ground.png 	new file:   assets/projectiles/golf.gif 	new file:   assets/projectiles/golf_explosion.gif 	renamed:    assets/image/chara_attack.gif -> assets/trump/attack.gif 	renamed:    assets/sound/atk_sound.wav -> assets/trump/attack_0.wav 	new file:   assets/trump/attack_1.wav 	new file:   assets/trump/attack_2.wav 	new file:   assets/trump/attack_3.wav 	new file:   assets/trump/hurt.gif 	renamed:    assets/image/chara_idle.gif -> assets/trump/idle.gif 	renamed:    assets/image/chara_run.gif -> assets/trump/run.gif 	new file:   character/guodong.c 	new file:   character/guodong.h 	new file:   character/jinping.c 	new file:   character/jinping.h 	new file:   character/trump.c 	new file:   character/trump.h 	modified:   element/character.c 	modified:   element/character.h 	modified:   element/ground.c 	modified:   element/ground.h 	modified:   element/projectile.c 	modified:   element/projectile.h 	deleted:    element/teleport.c 	deleted:    element/teleport.h 	deleted:    element/tree.c 	deleted:    element/tree.h 	modified:   game.exe 	modified:   physics/collision.c 	modified:   physics/collision.h 	modified:   physics/contact.c 	modified:   physics/gravity.c 	modified:   physics/gravity.h 	modified:   physics/physics.c 	modified:   physics/physics.h 	new file:   projectiles/golf.c 	new file:   projectiles/golf.h 	new file:   projectiles/golf_explosion.c 	new file:   projectiles/golf_explosion.h 	new file:   projectiles/twitter.c 	new file:   projectiles/twitter.h 	new file:   projectiles/twitter_explosion.c 	new file:   projectiles/twitter_explosion.h 	modified:   scene/gamescene.c 	modified:   scene/gamescene.h 	modified:   shapes/Circle.c 	modified:   shapes/Circle.h 	modified:   shapes/Rectangle.h 	new file:   tree.o

+++ physics/contact.c
@@ -3,10 +3,10 @@
 void Contact_Compute_Hitbox(Shape *hitbox, ContactInfo *info, ElementVec *ground_elements) {
     if (!hitbox || !info || !ground_elements) return;
 
-    double top    = hitbox->get_top(hitbox);
-    double bottom = hitbox->get_bottom(hitbox);
-    double left   = hitbox->get_left(hitbox);
-    double right  = hitbox->get_right(hitbox);
+    float top    = hitbox->get_top(hitbox);
+    float bottom = hitbox->get_bottom(hitbox);
+    float left   = hitbox->get_left(hitbox);
+    float right  = hitbox->get_right(hitbox);
 
     info->wall_top_y = -1;
     info->wall_bot_y = -1;
